+++ mlfe/unsupported/nodes/initializer_f_cuda.cpp
@@ -12,7 +12,7 @@ namespace mlfe { namespace node {
 template <typename T, typename D = CUDAContext>
 struct ConstantInitCudaF : NodeFunctor {
     void Init(OperatorContext *oc) override{
-        _val = oc->attr->GetParam<int>("Value");
+        _val = oc->attr->GetParam<double>("Value");
         _x = oc->inputs[0];
     }
